% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPAQ_Shiny.R

\name{GPAQ_Shiny}
\alias{GPAQ_Shiny}
\title{GPAQ_Shiny: Shiny app for the Global Physical Activity Questionnaire (GPAQ)}
\usage{
GPAQ_Shiny()
}

\description{
Lunch a shiny app on the local server for weighted and unweighted analysis
}

\examples{
GPAQ_Shiny()

}
© 2022 Git
