% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PtotalCat_svy_mean.R
\name{PtotalCat_svy_mean}
\alias{PtotalCat_svy_mean}
\title{PtotalCat_svy_mean: Calculating the Weighted Mean for the Total Physical Activity Categories and any Combination of the Factors for the Global Physical Activity Questionnaire (GPAQ)}
\usage{
PtotalCat_svy_mean(Outcome, Group = NULL, Data, id, weights, strata, CLN)
}
\arguments{
\item{Outcome}{the outcome of interest. PtotalCat is the outcome of interest with three levels "Low", "Moderate", and "High".}

\item{Group}{the group of interest. The default is NULL for when only the outcome variable is in our interest. Group can be any combination of categorical variables. Examples for group can be: ~age4y, ~age4y+sex, ~age4y+UrbanRural+sex, etc.}

\item{Data}{the output object of gpaq function.}

\item{id}{an argument in the "svydesign" function.}

\item{weights}{an argument in the "svydesign" function.}

\item{starta}{an argument in the "svydesign" function.}

\item{CLN}{is the "cln_ptotal" for the analysis of the outcome variable "ptotalCat".}

}
\value{
Output is a matrix with number of rows equal to the combination number of categories for Group arguments and seven columns. The first column indicate the group of interest, columns two to seven are the average and 95 percent confidence intervals for Low, Moderate, and High categories for total physical activity, respectively.
}
\description{
Prepare the output for Total Physical Activity Categories (Low, Medium, and High) explained in the GPAQ Analysis Guide at the population level
}
\examples{
   PtotalCat_svy_mean(~ptotalCat,~age4y,Data=Data,id=psu, weights=wstep1,strata =stratum,CLN="cln_ptotal")
   PtotalCat_svy_mean(~ptotalCat,~age4y+sex,Data=Data,id=psu, weights=wstep1,strata =stratum,CLN="cln_ptotal")
   PtotalCat_svy_mean(~ptotalCat,~age4y+ur+sex,Data=Data,id=psu, weights=wstep1,strata =stratum,CLN="cln_ptotal")

   PtotalCat_svy_mean(~ptotalCat,Data=Data,id=psu,	weights=wstep1,strata =stratum,CLN="cln_ptotal")

}
